pip_requirements.txt
affine==2.3.1
anyio==3.6.1
asgiref==3.5.2
aspose-words==22.6.0
async-generator==1.10
attrs==21.4.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
earthengine-api==0.1.312
fastapi==0.78.0
Fiona==1.8.21
fonttools==4.33.3
future==0.18.2
geojson==2.5.0
geomet==0.3.0
geopandas==0.10.2
h11==0.12.0
html2text==2020.1.16
httpcore==0.15.0
httplib2==0.20.4
httplib2shim==0.0.3
httpx==0.23.0
idna==3.3
imagecodecs==2022.2.22
imageio==2.19.3
iniconfig==1.1.1
kiwisolver==1.4.2
loguru==0.6.0
matplotlib==3.5.2
munch==2.5.0
networkx==2.8.3
numpy==1.22.4
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pyproj==3.3.1
pytest==7.1.2
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
pyvips==2.2.1
PyWavelets==1.3.0
rasterio==1.3a3
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
sentinelsat==1.1.1
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
snuggs==1.4.7
sortedcontainers==2.4.0
starlette==0.19.1
tifffile==2022.5.4
tomli==2.0.1
tqdm==4.64.0
trio==0.21.0
typing_extensions==4.2.0
uritemplate==3.0.1
urllib3==1.26.9
uvicorn==0.17.6
